+++ src/@next/components/organisms/ProductVariantPicker/fixtures.ts
@@ -6,7 +6,7 @@ export const productVariants: ProductDetails_product_variants[] = [
     id: "UHJvZHVjdFZhcmlhbnQ6MjAy",
     sku: "93855755",
     name: "1l / cotton / yellow",
-    stockQuantity: 50,
+    quantityAvailable: 50,
     isAvailable: true,
     images: [],
     pricing: {
@@ -98,7 +98,7 @@ export const productVariants: ProductDetails_product_variants[] = [
     id: "UHJvZHVjdFZhcmlhbnQ6MjA0",
     sku: "80884671",
     name: "2l / wool / red",
-    stockQuantity: 50,
+    quantityAvailable: 50,
     isAvailable: true,
     images: [],
     pricing: {
@@ -190,7 +190,7 @@ export const productVariants: ProductDetails_product_variants[] = [
     id: "UHJvZHVjdFZhcmlhbnQ6MzE2",
     sku: "100",
     name: "500ml / wool / turquoise",
-    stockQuantity: 50,
+    quantityAvailable: 50,
     isAvailable: true,
     images: [],
     pricing: {