+++ src/@sdk/fragments/gqlTypes/Checkout.ts
@@ -458,7 +458,7 @@ export interface Checkout_lines_variant {
   /**
    * Quantity of a product available for sale.
    */
-  stockQuantity: number;
+  quantityAvailable: number;
   /**
    * Whether the variant is in stock and visible or not.
    */