+++ src/@sdk/fragments/gqlTypes/CheckoutLine.ts
@@ -222,9 +222,9 @@ export interface CheckoutLine_variant {
   name: string;
   sku: string;
   /**
-   * Quantity of a product available for sale.
+   * Quantity of a product available for sale in one checkout.
    */
-  stockQuantity: number;
+  quantityAvailable: number;
   /**
    * Whether the variant is in stock and visible or not.
    */