+++ src/@sdk/fragments/products.ts
@@ -35,7 +35,7 @@ export const productVariantFragment = gql`
     id
     sku
     name
-    stockQuantity
+    quantityAvailable(countryCode: $countryCode)
     isAvailable
     images {
       id