+++ src/@sdk/mutations/gqlTypes/RemoveCheckoutPromoCode.ts
@@ -458,9 +458,9 @@ export interface RemoveCheckoutPromoCode_checkoutRemovePromoCode_checkout_lines_
   name: string;
   sku: string;
   /**
-   * Quantity of a product available for sale.
+   * Quantity of a product available for sale in one checkout.
    */
-  stockQuantity: number;
+  quantityAvailable: number;
   /**
    * Whether the variant is in stock and visible or not.
    */