+++ src/@sdk/mutations/gqlTypes/UpdateCheckoutShippingMethod.ts
@@ -471,9 +471,9 @@ export interface UpdateCheckoutShippingMethod_checkoutShippingMethodUpdate_check
   name: string;
   sku: string;
   /**
-   * Quantity of a product available for sale.
+   * Quantity of a product available for sale in one checkout.
    */
-  stockQuantity: number;
+  quantityAvailable: number;
   /**
    * Whether the variant is in stock and visible or not.
    */