+++ src/@sdk/queries/gqlTypes/CheckoutProductVariants.ts
@@ -2,6 +2,8 @@
 /* eslint-disable */
 // This file was automatically generated and should not be edited.
 
+import { CountryCode } from "./../../gqlTypes/globalTypes";
+
 // ====================================================
 // GraphQL query operation: CheckoutProductVariants
 // ====================================================
@@ -180,9 +182,9 @@ export interface CheckoutProductVariants_productVariants_edges_node_product {
 export interface CheckoutProductVariants_productVariants_edges_node {
   __typename: "ProductVariant";
   /**
-   * Quantity of a product available for sale.
+   * Quantity of a product available for sale in one checkout.
    */
-  stockQuantity: number;
+  quantityAvailable: number;
   /**
    * The ID of the object.
    */
@@ -226,4 +228,5 @@ export interface CheckoutProductVariants {
 
 export interface CheckoutProductVariantsVariables {
   ids?: (string | null)[] | null;
+  countryCode?: CountryCode | null;
 }