+++ src/userAccount/views/OrderDetails/gqlTypes/OrderByToken.ts
@@ -229,9 +229,9 @@ export interface OrderByToken_orderByToken_lines_variant {
   name: string;
   sku: string;
   /**
-   * Quantity of a product available for sale.
+   * Quantity of a product available for sale in one checkout.
    */
-  stockQuantity: number;
+  quantityAvailable: number;
   /**
    * Whether the variant is in stock and visible or not.
    */