+++ src/userAccount/views/OrderDetails/gqlTypes/ProductVariant.ts
@@ -188,7 +188,7 @@ export interface ProductVariant {
   /**
    * Quantity of a product available for sale.
    */
-  stockQuantity: number;
+  quantityAvailable: number;
   /**
    * Whether the variant is in stock and visible or not.
    */