+++ CHANGELOG.md
@@ -38,6 +38,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix unhandled JWT token expiration - #696 by @orzechdev
 - Regenerate types - #712 by @dominik-zeglen
 - Automatically choose variant attributes from url in ProductVariantPicker - #708 by @AlicjaSzu
+- Replace stockQuantity field with quantityAvailable - #723 by @AlicjaSzu
 
 ## 0.7.0